Corrected address casting (Mach-O part)

+++ src/parser/file/macho/cpp_optimized/machoFile.c
@@ -1,20 +1,22 @@
 /*
- * Callstack Library - Library creating human-readable call stacks.
+ * CallstackLibrary - Library creating human-readable call stacks.
  *
  * Copyright (C) 2023 - 2024  mhahnFr
  *
- * This file is part of the CallstackLibrary. This library is free software:
- * you can redistribute it and/or modify it under the terms of the
- * GNU General Public License as published by the Free Software Foundation,
- * either version 3 of the License, or (at your option) any later version.
+ * This file is part of the CallstackLibrary.
  *
- * This library is distributed in the hope that it will be useful,
+ * The CallstackLibrary is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * The CallstackLibrary is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
  *
- * You should have received a copy of the GNU General Public License along with
- * this library, see the file LICENSE.  If not, see <https://www.gnu.org/licenses/>.
+ * You should have received a copy of the GNU General Public License along with the
+ * CallstackLibrary, see the file LICENSE.  If not, see <https://www.gnu.org/licenses/>.
  */
 
 #include "../machoFile.h"
@@ -65,7 +67,7 @@ void machoFile_addFunction(struct machoFile* me, pair_funcFile_t function) {
 optional_debugInfo_t machoFile_getDebugInfo(struct machoFile* me, void* address) {
     struct machoFile_private* self = me->priv;
     
-    const uint64_t searchAddress = (uint64_t) (address - self->_._.startAddress) 
+    const uint64_t searchAddress = (uintptr_t) (address - self->_._.startAddress)
                                  + (me->inMemory ? me->text_vmaddr : self->_.addressOffset);
     
     pair_funcFile_t* closest = NULL;
